requirements.txt
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
coverage==7.10.7
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
execnet==2.1.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pycryptodome==3.23.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-mock==3.15.1
pytest-xdist==3.8.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
